locked some commands to testers

+++ commands/ranking/remove_all_ranks.js
@@ -1,5 +1,6 @@
 const { SlashCommandBuilder, EmbedBuilder, PermissionsBitField } = require('discord.js');
 const db = require("../../dbObjects.js")
+const testers = require("../../tester_servers.json");
 
 
 module.exports = {
@@ -11,8 +12,19 @@ module.exports = {
 
     async execute(interaction) {
         await interaction.deferReply()
-            const removeCount = await db.Ranks.destroy({where: { guild_id: interaction.guild.id}})
-            const embededReply = new EmbedBuilder().setColor([255, 255, 0]).setDescription(`successfuly removed all ${removeCount} ranks`)
-            interaction.editReply({embeds: [embededReply]})
+        const embeded_error = new EmbedBuilder().setColor([255,0,0])
+        let tester = false
+        testers.servers.forEach(server => {
+            if ( !tester && server.id === interaction.guild.id) {
+                tester = true
+            }
+        });
+        if (!tester) {
+            return await interaction.editReply({ embeds: [embeded_error.setDescription('This command is **only enabled** for testers!')] });
+        }  
+
+        const removeCount = await db.Ranks.destroy({where: { guild_id: interaction.guild.id}})
+        const embededReply = new EmbedBuilder().setColor([255, 255, 0]).setDescription(`successfuly removed all ${removeCount} ranks`)
+        interaction.editReply({embeds: [embededReply]})
     }
 }
